- coloquei um icone na toolbar para fazer a visualizacao; - agora a lista soh eh preenchida com updates realmente novos; - mudei a visualizacao de timeline, ficou bem mais legal, com uma barra horizontal tambem. Ainda falta arrumar para o caso de ter updates para um unico dia (caso dos 20 updates para o user logado. Dar um jeito de inserir mais um dia em branco para consertar isto; - colocado UserDeepT e StatusDeepT onde achei necessario; - No timeline as formas representam um StatusDeepT de onde as infos (como categoria) serao recolhidas para visualização.
Próximos passos:
- mostrar as infos no timeline;
- categorias;
- padronizar toolbar para que os botoes soh sejam adicionados para a tab a que pertencem;
- implementar busca de texto na toolbar para os updates (dixar pro final);
- mensagens diretas;
- menu do botao direito no user da rede nao deve conter acoes nao permitidas para aquele usuario;

+++ twitter4j/src/main/java/twitter4j/Twitter.java
@@ -593,13 +593,13 @@ public class Twitter implements java.io.Serializable {
                 "since_id", String.valueOf(since_id), true).asDocument(), this);
     }
     
-    public List<Status> getUserTimelineBySinceId(String id, int sinceID, int count) throws TwitterException{
+    public List<Status> getUserTimelineBySinceId(String id, long sinceID, int count) throws TwitterException{
     	return Status.constructStatuses(get(baseURL + "statuses/user_timeline/"+id+".xml",
     			"since_id", String.valueOf(sinceID), "count", String.valueOf(count), true).asDocument(), this);
     	
     }
     
-    public List<Status> getUserTimelineBySinceId(String id, int sinceID) throws TwitterException{
+    public List<Status> getUserTimelineBySinceId(String id, long sinceID) throws TwitterException{
     	return Status.constructStatuses(get(baseURL + "statuses/user_timeline/"+id+".xml",
     			"since_id", String.valueOf(sinceID), true).asDocument(), this);
